Cleans up extra fields apis and implements local field redaction
This addresses all issues mentioned here, except the decode side of
BaggageCodec, which is just saved for a different PR.

https://github.com/openzipkin/brave/pull/1182#issuecomment-621575987

Notably, this decouples `Extra` from `ExtraBaggageFields` even more
cleanly than before. This allows tests to be focused and also allows us
to reuse the base infrastructure for sampling as we did before.

Instead of having two data structure implementations of Baggage, this
uses one. The only difference between static and dynamic is that the
latter is allowed to add up to 64 more fields that aren't in the
whitelist.

This fixes the problem where local fields were leaked to the Baggage
encoder. This is implemented by a filtered map.

All baggage state is now implemented by `UnsafeArrayMap` which is a
view over a pair-wise array of baggage fields and potentially null
values.